First SQL project
This project aims to explore Walmart sales data to understand top performing branches and products, sales trend of different products and customer behavior. This dataset was obtained from Kaggle Walmart sales forecasting competition. The purpose of this project is to gain understanding from Walmart sales data exploring the various factors that influence across different branches. I started the project by creating a database called "salesdatawalmart"  and it was set as the default database in order to avoid referencing the database multiple times in the course of the project. I created a table "sales" which consists of 17 columns in the database then I imported the data into MySQL workbench  from an excel sheet. I also ensured that there were no null values in my table by adding the "not null" constraint for each column. This stage of the project is known as data wrangling.
The project had many other parts which are as follows:
1) Feature Engineering 
2) Exploratory Data Analysis
-Generic Questions
-Product Analysis
-Sales Analysis
-Customer Analysis