2nd SQL Project
Faasos is an online food ordering company. Faasos specializes in two different types of rolls which are veg roles and non veg roles. The datasets used in this video covers faasos customer orders, the ingredients used in each order ,the type of roll(s) ordered, the pick up time by each driver and so on. This project consists of different metrics and each metric has a bunch of questions. The metrics are as follows:
1) Roll Metrics 
Some of the questions answered in this metric are as follows:
-How many rolls were ordered?
-How many unique customers were made?
-How many successful orders were made by each driver?
2) Driver and Customer Experience
Some of the questions answered in this metric are as follows
-What is the average time in minutes it took each driver to arrive at the headquarters to pick up the order?
-Is there any relationship between number of rolls and how long the order takes to prepare?
-What is the average distance travelled for each customer?