Third SQL project
This project is all about pizza sales. The dataset for this project contains information such as pizza_id, pizza_name, pizza_ingredients, pizza_size ,unit_price and so on. The first thing that I did in this project was to create a schema called "pizzaDB" and it was set as the default schema. The dataset of this project was in an excel file, then I imported the data into MySQL workbench. After installing the dataset, I did some modifications to the data such as changing the datatypes of some columns. After the modifications, I carried out Exploratory Data Analysis to answer some questions.
The questions answered are as follows:
1) What is the total revenue?
2) Average amount spent per order
3) Total pizza sold
4) Total orders
5) Average amount of pizza per order
Questions Answered Based on Chart Requirements
1) Daily trend for total orders
2) Monthly trend for total orders
3) Percentage of sales by pizza category
4) Percentage of sales by pizza size
5) Top five best and worst sellers by revenue 
6) Top five best and worst sellers by quantity
7) Top five best and worst sellers by orders