Fourth SQL Project
This project is all about designing and building a SQL database. This is the project scenario:
The client Ben is opening a pizzeria in his town. The pizzeria won't be a dining, it would be just take out and delivery. The client has given us a project brief and the brief requires us to build a tailor made relational database that would allow him to capture and store all of the important information that the business generates. This would in turn help our client Ben, to monitor the performance of the business. There are three main areas which the brief requires us to concentrate on which are customer orders, stock control and staff. The way this task of designing the database and the tables in it is as follows:
1) The database would be created 
2) Each table and the respective fields would be created within the database 
3) A normalization process would be carried out on the data requiring us to add more related tables and defining the table relationships with the use of foreign keys.
A foreign key is a field in a table which is also the field with the primary key in another table. In this project, I was faced with a situation whereby the two fields involved to create the relationship between the two tables were not the primary keys in their respective tables. What I did in this this situation was to add a unique constraint to the field in the parent table for the relationship between the two tables to occur.