First Excel Project
This project is the same as the pizza_sales project in the main branch. The first one was done in SQL and I did this one in excel. Since I had the dataset within a table in one of the databases in MySQL workbench, I used MySQL connector/ODBC to export the data from MySQL workbench into excel by establishing a connection with the required database "pizzadb", then I selected the established connection via power query in excel and I loaded the data into my worksheet in excel. I added some new columns to the dataset such as "Day name" and "Month name", and I also wrote the name of the pizza_size in full by using the find and replace feature in excel.
After all the transformations to the dataset, I loaded the data into the power pivot window. I did this to calculate the measures that the project required me to calculate. Measures like Total revenue, Average amount per order, Total pizza sold, Total orders ,Average pizza per order. After this part of the project, I started answering the questions in the project with the use of pivot tables, then I proceeded to the chart requirements where I still used pivot tables to answer the necessary questions while getting the required pivot chart simultaneously. 
When I was done with analysis, I created two different dashboards which showed the results of the analysis carried out in the chart requirement section of the project.