Fifth SQL Project
This Project is about indian census. The dataset(s) of this project was in an Excel sheet. The first thing I did was to create a database called indiancensus and I set
it as the default schema.Then I imported the datasets from MySQL Workbench. The second dataset called new_populationdata contains columns such as district,
state,population and so on. The first dataset called population_data had columns such as District,state,growth,sex_ratio and so on. After importing the data into 
MySQL Workbench,the next step in the project was Exploratory data analysis. Some of the questions answered are:
1)Total population of India.
2)Average population growth in India.
3)Average population growth per state in India.
4)Average sex-ratio in India.
5)Average literacy rate. 
6)Average literacy rate greater than 90.
7)Three states with the highest average growth ratio.
8)Three states with the least average sex-ratio.
9)Number of males and females per state.
10)States starting with the letter "a" or "b".